Qbkm

+++ Components/hardware/inc/mpu6050.h
@@ -62,22 +62,22 @@ typedef struct
     int16_t Accel_X_RAW;
     int16_t Accel_Y_RAW;
     int16_t Accel_Z_RAW;
-    double Ax;
-    double Ay;
-    double Az;
+    float Ax;
+    float Ay;
+    float Az;
 
     int16_t Gyro_X_RAW;
     int16_t Gyro_Y_RAW;
     int16_t Gyro_Z_RAW;
-    double Gx;
-    double Gy;
-    double Gz;
+    float Gx;
+    float Gy;
+    float Gz;
 
     int16_t Temperature_RAW;
     float Temperature;
 
-    double KalmanAngleX;
-    double KalmanAngleY;
+    float KalmanAngleX;
+    float KalmanAngleY;
 
     MPU6050_config_t config;
